+++ CHANGELOG.md
@@ -383,3 +383,4 @@ Magento CE 1.8+ or EE 1.13+, see [these instructions](https://github.com/nexcess
   * [#739] Ban CMS Page after a CMS Revision has been updated (@ashsmith)
   * Added fix for community poll problems
   * [#516] Fixes multiple sessions generated on first page visit (@jharrisonau)
+  * [#796] Ignore additional GET Parameters in Varnish Cache via Configuration (@thampe)